Add DeepSeek-V4 decode sparse attention with grouped output projection
